+++ NEWS.md
@@ -1,3 +1,5 @@
+# vitals (development version)
+
 # vitals 0.2.0
 
 ## New features