+++ NEWS.md
@@ -1,5 +1,8 @@
 # vitals (development version)
 
+* Updated the vendored Inspect Log Viewer to Inspect version 0.3.151 
+  (released 30 November 2025, #194).
+
 # vitals 0.2.0
 
 ## New features