+++ NEWS.md
@@ -4,8 +4,8 @@
   to extract structured data from model responses. `generate_structured()` is to
   `parallel_chat_structured()` as `generate()` is to `parallel_chat()` (#153).
 
-* Updated the vendored Inspect Log Viewer to Inspect version 0.3.151
-  (released 30 November 2025, #194).
+* Updated the vendored Inspect Log Viewer to Inspect version 0.3.161
+  (released 14 January 2026, #194).
 
 * The log viewer home page will now show full summaries of each eval.