+++ NEWS.md
@@ -1,4 +1,4 @@
-# vitals (development version)
+# vitals 0.3.0
 
 ## New features